[renderer][main] refactor to independent from lastAction
