--- src/renderer/views/setting/DecibelDividerSetting.tsx
@@ -1,28 +1,0 @@
-/*
-Copyright 2024 Voibo
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-export type DecibelDividerSettingState = {
-  silenceLevel: number; //db
-  silenceLength: number; // sec
-  maxLength: number; // sec
-  minLength: number;
-};
-
-export const DecibelDividerSettingDefault: DecibelDividerSettingState = {
-  silenceLevel: -60, //db
-  silenceLength: 0.5, // sec
-  maxLength: 30, // sec
-  minLength: 5, // sec
-};
