+++ src/renderer/views/store/useVBStore.tsx
@@ -15,20 +15,11 @@ limitations under the License.
 */
 import { create } from "zustand";
 import { subscribeWithSelector } from "zustand/middleware";
-import { Segment } from "../../../common/Segment.js";
-import { AIConfig } from "../common/aiConfig.js";
-import { DiscussionSegment } from "../discussion/DiscussionSegment.js";
-import { DiscussionSplitterConf } from "../topic/DiscussionSplitter.js";
-import { Topic } from "../topic/Topic.js";
-import { VirtualAssistantConf } from "./useAssistantsStore.js";
-import {
-  NO_MINUTES_START_TIMESTAMP,
-  useMinutesStore,
-} from "./useMinutesStore.jsx";
-import { makeDefaultTitle, useMinutesTitleStore } from "./useMinutesTitle.jsx";
-import { IPCSenderKeys } from "../../../common/constants.js";
+import { Segment } from "../../../common/Segment.jsx";
 
 // ==== VB Core ====
+
+export const NO_MINUTES_START_TIMESTAMP = 0; // MinutesState
 
 export type VBStateMode = "home" | "recordingStudio";
 
@@ -50,11 +41,11 @@ export type VBState = {
   // == Topic ==
   interimSegment: Segment | null;
   // == zustand subscript ==
-  lastAction: VBAction | null;
+  //lastAction: VBAction | null;
 };
 
 type VBDispatch = {
-  vbDispatch: (action: VBAction) => void;
+  //vbDispatch: (action: VBAction) => void;
   isNoMinutes: () => boolean;
 };
 
@@ -76,335 +67,9 @@ export const useVBStore = create<VBState & VBDispatch>()(
     // == Topic ==
     interimSegment: null,
     // == zustand subscript ==
-    lastAction: null,
+    //lastAction: null,
 
-    vbDispatch: (action) => VBActionProcessor(action),
+    //vbDispatch: (action) => VBActionProcessor(action),
     isNoMinutes: () => get().startTimestamp === NO_MINUTES_START_TIMESTAMP,
   }))
 );
-
-export type VBAction =
-  | {
-      type: "createNewMinutes";
-    }
-  | {
-      type: "setMinutesLines";
-      payload: {
-        minutes: DiscussionSegment[];
-      };
-    }
-  | {
-      type: "setAudioFolder";
-      payload: {
-        audioFolder: string;
-      };
-    }
-  | {
-      type: "openMinutes";
-      payload: {
-        startTimestamp: number;
-      };
-    }
-  | {
-      type: "changeVADDialogOpen";
-    }
-  | {
-      type: "openHomeMenu";
-    }
-  | {
-      type: "deleteMinutes";
-      payload: {
-        startTimestamp: number;
-      };
-    }
-  | {
-      type: "setTopic";
-      payload: {
-        topics: Topic[];
-      };
-    }
-  | {
-      type: "removeTopic";
-      payload: {
-        topicID: string;
-      };
-    }
-  | {
-      type: "updateMinutesText";
-      payload: {
-        segmentIndex: number;
-        segmentTextIndex: number;
-        content: string;
-      };
-    }
-  | {
-      type: "removeMinutesText";
-      payload: {
-        segmentIndex: number;
-        segmentTextIndex: number;
-      };
-    }
-  | {
-      type: "splitMinutesText";
-      payload: {
-        segmentIndex: number;
-        segmentTextIndex: number;
-      };
-    }
-  | {
-      type: "mergeUpMinutesText";
-      payload: {
-        segmentIndex: number;
-        segmentTextIndex: number;
-      };
-    }
-  | {
-      type: "changeTopicStartedPoint";
-      payload: {
-        segmentIndex: number;
-      };
-    }
-  | {
-      type: "changeTopicAIConfig";
-      payload: {
-        aiConfig: AIConfig;
-      };
-    }
-  | {
-      type: "togglePlayWavMute";
-    }
-  | {
-      type: "deleteAllTopic";
-    }
-  | {
-      type: "addVirtualAssistantConf";
-      payload: {
-        assistant: VirtualAssistantConf;
-      };
-    }
-  | {
-      type: "setVirtualAssistantConf";
-      payload: {
-        assistant: VirtualAssistantConf;
-      };
-    }
-  | {
-      type: "removeVirtualAssistantConf";
-      payload: {
-        assistantId: string;
-      };
-    }
-  | {
-      type: "updateInterimSegment";
-      payload: {
-        segment: Segment;
-      };
-    }
-  | {
-      type: "changeDiscussionSplitterConf";
-      payload: {
-        splitterConf: DiscussionSplitterConf;
-      };
-    }
-  | {
-      type: "updateTopic";
-      payload: {
-        topic: Topic;
-      };
-    };
-
-export const VBActionProcessor = async (action: VBAction) => {
-  const minutesState = useMinutesStore(
-    useVBStore.getState().startTimestamp
-  ).getState();
-  switch (action.type) {
-    case "addVirtualAssistantConf":
-      minutesState.addVirtualAssistantConf(action.payload.assistant);
-      break;
-    case "removeVirtualAssistantConf":
-      minutesState.removeVirtualAssistantConf(action.payload.assistantId);
-      break;
-    case "setVirtualAssistantConf":
-      minutesState.setVirtualAssistantConf(action.payload.assistant);
-      break;
-    case "removeTopic":
-      minutesState.removeTopic(action.payload.topicID);
-      break;
-    case "deleteAllTopic":
-      minutesState.deleteAllTopic();
-      break;
-    case "changeTopicStartedPoint":
-      minutesState.changeTopicStartedPoint(action.payload.segmentIndex);
-
-      break;
-    case "changeTopicAIConfig":
-      minutesState.changeTopicAIConfig(action.payload.aiConfig);
-      break;
-    case "updateMinutesText":
-      minutesState.updateMinutesText(
-        action.payload.segmentIndex,
-        action.payload.segmentTextIndex,
-        action.payload.content
-      );
-      break;
-    case "removeMinutesText":
-      minutesState.removeMinutesText(
-        action.payload.segmentIndex,
-        action.payload.segmentTextIndex
-      );
-      break;
-    case "splitMinutesText":
-      minutesState.splitMinutesText(
-        action.payload.segmentIndex,
-        action.payload.segmentTextIndex
-      );
-      break;
-    case "mergeUpMinutesText":
-      minutesState.mergeUpMinutesText(
-        action.payload.segmentIndex,
-        action.payload.segmentTextIndex
-      );
-      break;
-    case "setTopic":
-      minutesState.setTopic(action.payload.topics);
-      break;
-    case "changeVADDialogOpen":
-      useVBStore.setState({
-        audioSettingsDialogOpen: !useVBStore().audioSettingsDialogOpen,
-      });
-      break;
-    case "setAudioFolder":
-      useVBStore.setState({
-        audioFolder: action.payload.audioFolder,
-      });
-      break;
-    case "setMinutesLines":
-      minutesState.setMinutesLines(action.payload.minutes);
-      useVBStore.setState({
-        interimSegment: null,
-      });
-      break;
-    case "togglePlayWavMute":
-      useVBStore.setState({
-        playWavMute: !useVBStore().playWavMute,
-      });
-      break;
-    case "updateInterimSegment":
-      useVBStore.setState({
-        interimSegment: action.payload.segment,
-      });
-      break;
-    case "changeDiscussionSplitterConf":
-      minutesState.changeDiscussionSplitterConf(action.payload.splitterConf);
-      break;
-    case "updateTopic":
-      minutesState.updateTopic(action.payload.topic);
-      break;
-
-    /** == MinutesDispatch関係 == */
-
-    case "createNewMinutes":
-      const startTimestamp = Date.now();
-      console.warn("useVBStore: createNewMinutes: 0", startTimestamp);
-      useMinutesStore(startTimestamp).getState().createNewMinutes();
-      // server
-      window.electron.send(IPCSenderKeys.CREATE_MINUTES, startTimestamp);
-      // title
-      useMinutesTitleStore.getState().setMinutesTitle({
-        title: makeDefaultTitle(startTimestamp),
-        startTimestamp: startTimestamp,
-      });
-      useVBStore.setState({
-        startTimestamp,
-        mode: "recordingStudio",
-        recording: false,
-        mainMenuOpen: false,
-      });
-      console.warn("useVBStore: createNewMinutes: 1", startTimestamp);
-      break;
-    case "openMinutes":
-      console.warn("useVBStore: openMinutes: 0", action.payload.startTimestamp);
-      await useMinutesStore(action.payload.startTimestamp)
-        .getState()
-        .waitForHydration()
-        .then((newState) => {
-          console.warn("useVBStore: openMinutes: 1", newState);
-        });
-      console.warn("useVBStore: openMinutes: 2", action.payload.startTimestamp);
-      useVBStore.setState({
-        startTimestamp: action.payload.startTimestamp,
-        mode: "recordingStudio",
-        mainMenuOpen: false,
-        recording: false,
-        interimSegment: null,
-      });
-      console.warn("useVBStore: openMinutes: 3", action.payload.startTimestamp);
-      break;
-    case "openHomeMenu":
-      console.log("useVBStore: openHomeMenu: 0");
-      await useMinutesStore(NO_MINUTES_START_TIMESTAMP)
-        .getState()
-        .waitForHydration()
-        .then((newState) => {
-          console.log("useVBStore: openHomeMenu: 1", newState);
-        });
-      console.log("useVBStore: openHomeMenu: 2");
-      useVBStore.setState({
-        startTimestamp: NO_MINUTES_START_TIMESTAMP,
-        mode: "home",
-        recording: false,
-        mainMenuOpen: true,
-        interimSegment: null,
-      });
-      console.log("useVBStore: openHomeMenu: 3");
-      break;
-    case "deleteMinutes":
-      console.log("useVBStore: deleteMinutes: 0");
-      minutesState.deleteMinutes();
-      await useMinutesStore(NO_MINUTES_START_TIMESTAMP)
-        .getState()
-        .waitForHydration()
-        .then((newState) => {
-          console.log("useVBStore: deleteMinutes: 1", newState);
-        });
-      console.log("useVBStore: deleteMinutes: 2");
-      useVBStore.setState({
-        startTimestamp: NO_MINUTES_START_TIMESTAMP,
-        mode: "home",
-        mainMenuOpen: true,
-        recording: false,
-      });
-      console.log("useVBStore: deleteMinutes: 3");
-      break;
-    default:
-      console.log("vbActionProcessor: unexpected default", action);
-  }
-  useVBStore.setState({
-    lastAction: action,
-  });
-};
-
-export function filterSequentialTopics(
-  topics: Topic[]
-): Array<{ index: number; topic: Topic }> {
-  const sequentialTopics: Array<{ index: number; topic: Topic }> = [];
-  let checked = true;
-  topics.forEach((topic) => {
-    if (!checked) return;
-    // index 確認
-    const targetIndex = useMinutesStore(useVBStore.getState().startTimestamp)
-      .getState()
-      .topics.indexOf(topic);
-    if (targetIndex === -1) {
-      checked = false;
-      return;
-    }
-    if (
-      sequentialTopics.length === 0 ||
-      sequentialTopics[sequentialTopics.length - 1].index + 1 === targetIndex
-    ) {
-      sequentialTopics.push({ topic, index: targetIndex });
-    }
-  });
-  return sequentialTopics;
-}
